refactor(cliente): modularizar el API de productos
