+++ client/src/services/api/productos/index.ts
@@ -1,0 +1,7 @@
+import { getProductos as _getProductos } from './productos/get-productos';
+
+export const getProductos = _getProductos;
+
+export default {
+  getProductos
+};
