+++ client/src/store/modules/autenticacion/index.ts
@@ -13,7 +13,8 @@ import {
   LOGOUT,
   OBTENER_USUARIO,
   LOGIN_CLIENTE,
-  LOGOUT_CLIENTE
+  LOGOUT_CLIENTE,
+  OBTENER_PRODUCTOS
 } from '@/store/types/acciones';
 import { NOMBRE_USUARIO } from '@/store/types/getters';
 import { SET_USUARIO, SET_ESTA_LOGUEADO } from '@/store/types/mutaciones';
@@ -22,11 +23,23 @@ import { Usuario } from '@/types/usuario-tipos';
 import { Module } from 'vuex';
 
 import { crearNotificacion } from '../notificaciones/crear-notificacion';
+import {
+  obtenerEspacioDeNombres,
+  MODULO_PRODUCTOS
+} from '@/store/types/modulos';
 
 interface EstadoAutenticacion {
   estaLogueado: boolean;
   usuario: Usuario | null;
 }
+
+const obtenerProductos = obtenerEspacioDeNombres([
+  MODULO_PRODUCTOS,
+  OBTENER_PRODUCTOS
+]);
+
+// Acciones
+const ACTUALIZAR_REGISTROS = 'actualizarRegistros';
 
 const moduloAutenticacion: Module<EstadoAutenticacion, EstadoBase> = {
   namespaced: true,
@@ -52,6 +65,7 @@ const moduloAutenticacion: Module<EstadoAutenticacion, EstadoBase> = {
         if (respuesta.ok) {
           await dispatch(OBTENER_USUARIO);
           commit(SET_ESTA_LOGUEADO, true);
+          await dispatch(ACTUALIZAR_REGISTROS);
         }
         return respuesta;
       } catch (error) {
@@ -77,6 +91,7 @@ const moduloAutenticacion: Module<EstadoAutenticacion, EstadoBase> = {
         if (respuesta.ok) {
           commit(SET_USUARIO, null);
           commit(SET_ESTA_LOGUEADO, false);
+          await dispatch(ACTUALIZAR_REGISTROS);
         }
 
         crearNotificacion(dispatch, respuesta);
@@ -85,6 +100,10 @@ const moduloAutenticacion: Module<EstadoAutenticacion, EstadoBase> = {
       } catch (error) {
         throw error;
       }
+    },
+
+    async [ACTUALIZAR_REGISTROS]({ dispatch }): Promise<any> {
+      return dispatch(obtenerProductos, null, { root: true });
     },
 
     [LOGIN_CLIENTE]({ commit }, usuario: Usuario): void {
